% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prs.R
\name{create_prs}
\alias{create_prs}
\title{Pruning+thresholding polygenic risk score creation based on summary statistics}
\usage{
create_prs(
  variant_data,
  gwas_info,
  remove_indels = FALSE,
  imp_threshold = 0.8,
  binary_outcome = TRUE,
  exclude_extreme_associations = TRUE,
  LDplot = FALSE,
  pruning_threshold = 0.75,
  pval_threshold = 5e-08,
  conditional = FALSE,
  cond_window = 1e+07,
  cond_N = 60000,
  scale = FALSE,
  flowchart = TRUE
)
}
\arguments{
\item{variant_data}{An object of format output by extract_variants().}

\item{gwas_info}{An object generated by get_trait_variants() or get_pQTLs().}

\item{remove_indels}{If TRUE, removes indels.}

\item{imp_threshold}{Imputation quality threshold, based on R^2. Any variant with lower imputation R^2 is removed.}

\item{binary_outcome}{Set to TRUE for binary traits, and FALSE for continuous outcomes (including pQTLs).}

\item{exclude_extreme_associations}{If TRUE, removes variants with an odds ratio > 5 or <1/5.}

\item{LDplot}{If TRUE, plots the LD matrix (squared correlation matrix of variants).}

\item{pruning_threshold}{Variants in LD >= pruning_threshold with other variants are removed, keeping higher MAF variants.}

\item{pval_threshold}{Variants with GWAS p-values > pval_threshold are discarded. Set to 1 to turn off.}

\item{conditional}{If TRUE, uses marg2con() to create decorrelate variants in LD if these are within a given bp distance on the same chromosome.}

\item{cond_window}{A genomic distance within which to decorrelate variants in LD, in base pair.}

\item{cond_N}{The sample size of the original GWAS from which the marginal estimates were derived, or an approximation of it.}

\item{scale}{Centers and standardizes the polygenic risk score if TRUE.}

\item{flowchart}{If TRUE, plots a flowchart describing the creation of the polygenic risk score.}
}
\value{
A list containing several data.frames with all relevant information. The risk score is stored in element 'prs'.
}
\description{
Create a polygenic risk score based on summary statistics from prior GWAS/pQTL discovery studies.
Included variants in high LD can be decorrelated to prevent double-counting, using the conditional argument.
}
\examples{
# vte_prs <- create_prs(vte_extracted_variants, vte_gwas_info)
# hist(vte_prs$prs$prs)
}
