% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pQTLs.R
\name{get_pQTLs}
\alias{get_pQTLs}
\title{Retrieve protein quantitative trait loci (pQTLs) with summary statistics}
\usage{
get_pQTLs(gene_symbol, risk_frequency = TRUE, is_cis = TRUE)
}
\arguments{
\item{gene_symbol}{Expects a character vector of length 1, containing the exact gene symbol.}

\item{risk_frequency}{If TRUE, also includes the frequency of the risk allele in the output.}

\item{is_cis}{If TRUE, only includes cis pQTLs in the output.}
}
\value{
A formatted data.frame containing all pQTLs for a given gene symbol, including summary statistics.
}
\description{
Returns a formatted data.frame of variants for a given gene symbol, including summary statistics
}
\examples{
ACE2_pqtls <- get_pQTLs('ACE2',pQTL_list, is_cis=FALSE)
}
