% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trait_variants.R
\name{get_trait_variants}
\alias{get_trait_variants}
\title{Retrieve trait-related variants and summary statistics from the GWAS catalog}
\usage{
get_trait_variants(
  trait,
  trait_list = NULL,
  risk_frequency = TRUE,
  mapped_gene = FALSE,
  ask_about_efo = TRUE
)
}
\arguments{
\item{trait}{Expects a regular expression to find the trait of interest}

\item{trait_list}{An object generated by init_gwas_db(). If NULL, init_gwas_db() will be run internally.}

\item{risk_frequency}{If TRUE, also includes the frequency of the risk allele in the output.}

\item{mapped_gene}{If TRUE, includes the mapped gene for each variant in the output.}

\item{ask_about_efo}{If TRUE, requires dynamic user input to specify which EFO id should be used.}
}
\value{
A formatted data.frame containing all trait-related variants, including summary statistics.
}
\description{
Returns a formatted data.frame of variants for a given trait, including summary statistics
}
\examples{
\donttest{MI_variants <- get_trait_variants('MI|myocardial infarction',trait_list=trait_list,ask_about_efo=FALSE)}
}
