% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pQTL-data.R
\docType{data}
\name{pQTL_list}
\alias{pQTL_list}
\title{Protein quantitative trait loci (pQTLs) for 1397 proteins, from Eldjarn et al. (bioRxiv 2022)}
\format{
An object of class \code{"data.frame"}.
}
\source{
\href{https://www.biorxiv.org/content/10.1101/2022.02.18.481034v1.abstract}{BioRxiv}
}
\usage{
data(pQTL_list)
}
\description{
Data from a pQTL discovery study by Eldjarn et al. (bioRxiv 2022), using ~50k individuals from the UK Biobank.
Protein levels were measured by proximity extension assay (Olink Explore 1536, Olink Proteomics, Uppsala, Sweden).
}
\examples{
data(pQTL_list)
str(pQTL_list)

}
\references{
Eldjarn GH, Ferkingstad E, Lund SH et al."Large-scale comparison of immunoassay and aptamer-based plasma proteomics
through genetics and disease." BioRxiv 2022
(\href{https://www.biorxiv.org/content/10.1101/2022.02.18.481034v1.abstract}{BioRxiv})
}
\keyword{datasets}
