% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_grid.R
\name{prs_grid}
\alias{prs_grid}
\title{Create a grid of polygenic risk scores with different tuning parameter combinations}
\usage{
prs_grid(
  variant_data,
  gwas_info,
  binary_outcome = TRUE,
  imp_range = seq(0.8, 1, 0.05),
  pruning_range = seq(0.7, 1, 0.05),
  pval_range = quantile(gwas_info$pvalue, seq(0.1, 1, 0.1)),
  optimal = FALSE,
  optimal_trials = 9
)
}
\arguments{
\item{variant_data}{An object of format output by extract_variants().}

\item{gwas_info}{An object generated by get_trait_variants() or get_pQTLs().}

\item{binary_outcome}{Set to TRUE for binary traits, and FALSE for continuous outcomes (including pQTLs)}

\item{imp_range}{A vector of imputation R^2 values.}

\item{pruning_range}{A vector of LD (R^2) pruning threshold values.}

\item{pval_range}{A vector of p-value thresholds.}

\item{optimal}{If TRUE, creates the grid according to a D-optimal fractional factorial design.}

\item{optimal_trials}{Number of rows in the fractional factorial design matrix.}
}
\value{
A list containing the grid design, a list of all results, and a data.frame with all scores.
# grid <- prs_grid(MI_variants, MI_gwas)
}
\description{
Create a full or fractional factorial grid of polygenic risk scores based on a
user-specified range of tuning parameter values. Useful to evaluate the influence of various
tuning parameters, and for the creation of stacked scores.
}
