% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_prs.R
\name{stacked_prs}
\alias{stacked_prs}
\title{Stacked P+T polygenic risk scores by regularized regression}
\usage{
stacked_prs(
  prs_grid,
  y,
  nfolds = 10,
  family = "binomial",
  plot_error = TRUE,
  seed = NULL,
  ...
)
}
\arguments{
\item{prs_grid}{The element 'result_df' from the list output by prs_grid().}

\item{y}{The observed outcome (trait).}

\item{nfolds}{Number of folds to use in k-fold cross-validation within cv.glmnet().}

\item{family}{The distribution family of the outcome, e.g. binomial, gaussian.}

\item{plot_error}{If TRUE, plots the error of the stacked PRS by the tuning parameter lambda.}

\item{seed}{If not NULL, sets a seed prior to running cv.glmnet().}

\item{...}{Additional arguments to pass to cv.glmnet().}
}
\value{
A list containing the model, the stacked PRS, and the non-zero model coefficients.
}
\description{
Uses individual-level data to combine ('stack') polygenic risk scores based on
different tuning parameter combinations to optimize predictive performance. Based on
Privé et al. AJHG 2019.
}
